doc/module-system: Introduce concepts
This is reference material in a form that is meant to provide
somewhat of an overview.

It doesn't defers to other sections to provide the full depth,
which is uncharacteristic of reference material.
It is however not an explanation, as it is does not cover the why;
only the what. (Except for the eDSL part I guess)