nixos/tpm2: fix unconfigured fapi-config.json
When TSS2_FAPICONF is not set, the fapi-config.json from the package is
used.

This causes failures due to attempted writes to the nix store as the
paths in that file have been substituted (profile_dir needs substitution
but system_dir and log_dir do not).

Additionally, the tcti is not configured in the default file, and should
be set to same value as TPM2TOOLS_TCTI/TPM2_PKCS11_TCTI.

Fixes https://github.com/NixOS/nixpkgs/issues/192771