darwin."10.13.6".bsdmake: Include from developer-tools-3.1.3
Current source releases get this from version 3.2.6 of the Developer
Tools. However, this seems arbitrary since they were available as recent
as version 4.2 and as early as version 3.1.3.

My rationale for picking the oldest available source is if we introduced
it as early as that version we wouldn't have needed to ever bump it.
It's also a more honest representation of how old the sources we depend
on are.