glibcHeaders: init at 2.37-8
This expression allows to build the glibc headers without first
having built a targetPlatform compiler.  It is needed in order to
break the dependency cycle between glibc and libgcc:

1. glibc must have the exact path to libgcc embedded into it via
   user-defined-trusted-dirs; see nixpkgs commit
   7553d0fe29801938bcb280bb324b579ef9016aea.  This reference is not
   part of the ELF structure so we can't use `patchelf` to update it.

2. libgcc must be able to see the glibc headers in order to compile
   unwinder support.  Otherwise, it will deliberately `assert()`
   when a thread which has called `pthread_cleanup_push()` exits:

   https://gcc.gnu.org/git/?p=gcc.git;a=blob;f=libgcc/unwind-dw2.c;h=d0afce7a9ea9f5b12a5a01ef1e940e1452b48cab;hb=HEAD#l1336

Unfortunately due to how GNU autoconf works, you can't *usually* get
the glibc headers without a targetPlatform compiler, because the
headers are partially generated based on autoconf results, and
autoconf is based on running the targetPlatform C compiler on
various inputs and seeing if they fail.

Fortunately the glibc developers are at least partially aware of
this being problematic, which is why they have avoided using any
targetPlatform-compiler-derived data when producing the
targetPlatform headers, and have provided the `--enable-hacker-mode`
flag for `./configure`, which runs just enough of the process to
allow `make install-headers` to work (but `make` certainly *won't*).