kanidm: apply migration fixes
It was noticed, that on a direct upgrade from alpha10 to beta12 a
partial migration would take place, which would fail, but not rollback
completely.

This would then prevent an upgrade from alpha10 to alpha11 and leave you
in an unusable state, requiring you to restore from a backup.

Upstream now quickly provided a maintenance branch for beta12, to which
we are forwarding here.

(cherry picked from commit 0896ea435b94082550cd769e097e9f9b152e2ef9)