passage: use getopt from nix on darwin
This removes an ugly hack from upstream which hard-codes a bunch of
paths to find gnu-getopt (as opposed to darwins getopt in
/usr/bin/getopt).

This lets nix-built passage fail on systems which don't have one of
those already installed (i.e. from homebrew).

As upstream seemingly doesn't provide a way to override this easily yet,
we do it ourselves.