wasilibc: allow enabling experimental threads support
This is an experimental new feature in wasilibc which also allows
enabling threads in libcxx* as a consequence. This is wired up using
passthru.hasThreads. It is disabled by default (as it is experimental
still), but overlaying wasilibc to turn on the flag should allow using
it.

See also:
- https://github.com/WebAssembly/wasi-sdk/pull/274
- https://github.com/WebAssembly/wasi-sdk/pull/301
- https://github.com/WebAssembly/wasi-sdk/pull/314

Flags for libc++abi are copied from wasi-sdk.

The BULK_MEMORY_SOURCES is dropped, as it no longer needs to be set to
be empty to compile firefox. With it set, the build of wasilibc would
fail if enableThreads is true.

Test using

  (import ./. {
    overlays = [
      (self: super: {
        wasilibc = super.wasilibc.override { enableThreads = true; };
      })
    ];
  }).wasilibc.tests.simple-c-cxx-binaries