stgit: Use phases from stdenv
Instead of overwriting the full build/check/install phases, use the phases from
stdenv. They allow to control the build by just setting a few variables for the
targets and a single one for the flags needed to be passed to make.
Additionally, they behave well with the rest of the build system, e.g., they
run the hooks correctly.

Also, the MAN_BASE_PATH is not used in the stgit build process and the
installed man pages work fine without it, so remove it.