nixos/paperless: fix json-encoded options
Some paperless options (like PAPERLESS_CONSUMER_IGNORES) are taken
as JSON-encoded strings. Using regular double-quotes and then using
\-quoted double quotes within the JSON string works for the systemd
unit but not for the manage command.

With this combination we can actually use `toJSON` to write those
values without quoting hassle as a regular nix expression.

Includes a test that verifies we're getting this right for both ends.

Thanks: erikarvsted for even better quoting.