python311Packages.python-arango: Add setuptools to propagatedBuildInputs
I was trying to use `python-arango` in one of my experiments with NixOS (23.05). I got the following error:

```console
[rm_me] λ nix-shell -p python311Packages.python-arango

[nix-shell:/data/rm_me]$ python -c 'import arango'
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/nix/store/4n18qmv3dfx8gja79rvrlmrxzv1acmx9-python3.11-python-arango-7.5.7/lib/python3.11/site-packages/arango/__init__.py", line 2, in <module>
    from arango.client import ArangoClient  # noqa: F401
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/nix/store/4n18qmv3dfx8gja79rvrlmrxzv1acmx9-python3.11-python-arango-7.5.7/lib/python3.11/site-packages/arango/client.py", line 6, in <module>
    from pkg_resources import get_distribution
ModuleNotFoundError: No module named 'pkg_resources'
```

To fix it, I have to pass an additional package to the nix shell:

```console
[rm_me] λ nix-shell -p python311Packages.python-arango python311Packages.setuptools

[nix-shell:/data/rm_me]$ python -c 'import arango'

[nix-shell:/data/rm_me]$
```

I found a similar issue with `pwndbg` in the past has been resolved in #71219. But I think in this case, we only need to change the Python side, not everything.

This is my first nixpkgs PR, apologies if it is not correct. Will try to make it work :)

Signed-off-by: Mihai Maruseac <mihai.maruseac@gmail.com>