writeTextFile: support in derivationArgs pre/postInstall and installCheckPhase
Support preInstall and postInstall in derivationArgs to specify commands
to run before/after target text file generation ("installation").

Support doInstallCheck and installCheckPhase in derivationArgs to check
the generated file.

Keep backward compatibility to checkPhase while print out warning
messages.

The same changes are applied to writeShellApplication, which is built on
top of writeTextFile. In addition, writeShellApplication now run
preInstallCheck/postInstallCheck hooks instead of preCheck/postCheck
hooks at the beginning and the end of the tests.