ruby: fix build with clang 16
Ruby requires `__cospi` and `__sinpi` from `math.h`. These should be
available in 10.12, but they’re not because the upstream header used by
the source-based SDK is from 2002. Unfortunately, there is no newer
version of that header available upstream, so the only way to fix the
calls to undefined library functions is by using the 11.0 SDK. These
worked on clang 11, but they are an error on clang 16.