chromium: supply Rust compiler for M115+
Chromium has recently started integrating Rust [1] and as of M115 this
infrastructure is enabled, thus Chromium either requires the bundled
binary Rust compiler (which won't work in NixOS without modification) or
a custom one. This configures the nixpkgs Rust compiler for Chromium.

M115 doesn't actually contain any production Rust code, so we don't
actually know if this fully works and we're definitely still missing
bindgen, but for the time being this makes it work and is a sane
starting position.

[1] https://security.googleblog.com/2023/01/supporting-use-of-rust-in-chromium.html