mold: patch an upstream bug in --strip-debug
Mold currently has a bug where `--strip-debug` is treated the same as
`--strip-all`; this causes things like Rust backtraces to stop working
because the symbol tables get stripped.

The fix is trivial and obvious, and comes from an upstream patch that
needs to be merged and release.

GitHub Issue: rui314/mold#1038
Signed-off-by: Austin Seipp <aseipp@pobox.com>