rke2: fix embedded versions
Use the provided scripts to build the project, in order to minimize the
differences between the Nix and the upstream distribution.

Among other things, our build wasn't embedding the docker image tags in
the binary, and that resulted in the binary trying to load "latest"
images, failing in the process.