lvm2: fix libdevmapper SONAME when onlyLib is on
Building cryptsetup with onlyLib on leads to linker errors, because the
SONAME of libdevmapper includes a version number, and the library is
installed without one.

This patch fixes that by delegating more of the install to make. That
uncovered another issue, where invalid placeholder strings were being
passed to the build, so I fixed that as well.

Tested on x86_64-linux by building .#cryptsetup, which succeeds. The
shared libraries are identical, but the new one is now a symlink to
libdevmapper.so.1.02. The store paths in the pkgconfig file also changed
(as expected).

After building .#pkgsStatic.cryptsetup, the only difference are the
paths in pkgconfig file (one of which was a placeholder anyway).