jujutsu: overhaul build expression
The `jujutsu` expression has changed very lightly in the past year since
it was introduced, but the upstream Nix flake expression has gotten a
lot of updates in that time, to make builds faster and more featureful.

This is a port of some of those upstream changes. Among other things,
this:

- Stops zstd from being vendored via crate
- Stops libgit2 from being vendored via crate
- Removes a spurious dbus(?) dependency
- Installs man pages
- Runs the testsuite via nextest

Signed-off-by: Austin Seipp <aseipp@pobox.com>