alsa-scarlett-gui: Fix
Attemt to fix the desktop integration for alsa-scarlett-gui as it would
not show up (at least in Gnome) in any menu. For the fix I added a patch
file that changes `Exec = PREFIX/bin/alsa-scarlett-gui` in the resulting
.desktop file to the two
lines:
```
Exec = alsa-scarlett-gui %F
TryExec = alsa-scarlett-gui
```
This way you must only make sure the executable file is in your PATH
environment variable.