microsoft-edge: ensure stable order for upstream sources
This commit addresses the issue of unstable key ordering in default.nix.
Previously, the keys could appear in different orders such as "stable,
beta, dev" or "beta, stable, dev".
To simplify the reviewing process and minimize merge conflicts, this
commit introduces the use of OrderedDict, ensuring consistent key
ordering for upstream sources.
Also, the gzip import was unused, so just drop it.