linux/generic: take a kernelTarget argument and deprecates legacy install targets
Nowadays, most architectures supports `KBUILD_IMAGE` which is
set by the architecture-specific Makefile or can be overriden by the user
Before that, we were using Makefile "install targets" such as "uinstall" for uImage,
"zinstall" for zImage, etc.

We move them into a "legacy" list which is used by default if `kernelTarget` is null.