lib.systems.supported: Add list of supported platforms
RFC #112 raised the question of what the source of truth is about the
platforms supported in Nixpkgs. RFC #46 specified the authoritative
source would be the Nixpkgs manual but this hasn't happened yet.

When it comes to Nixpkgs the tier list is currently only important to
determine what platforms Hydra should evaluate for. This information is
currently in `lib/systems/flake-systems.nix`. However, a manual section
as described in RFC #46 would be a second consumer, so a separate file
is warranted.

We re-introduce `lib/systems/supported.nix` as a suitable authoritative
source. A Nixpkgs manual section could be auto-generated based on this
file.